requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
boto3==1.24.0
psycopg2-binary==2.9.9
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
playwright==1.49.0
